Implemetacion de CustomException, y eliminacion de ExceptionDAO y ExceptionService

+++ src/main/java/com/blogPersonal/service/ComentarioServiceImp.java
@@ -2,8 +2,8 @@ package com.blogPersonal.service;
 
 import com.blogPersonal.dao.ComentarioDAO;
 import com.blogPersonal.dto.Comentario;
-import com.blogPersonal.exception.ExceptionDAO;
-import com.blogPersonal.exception.ExceptionService;
+import com.blogPersonal.exception.CustomException;
+import com.blogPersonal.util.MessagesConstant;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,61 +20,76 @@ public class ComentarioServiceImp implements ComentarioService {
     }
 
     @Override
-    public List<Comentario> findAll() throws ExceptionService {
+    public List<Comentario> findAll() throws CustomException {
         List<Comentario> comentarios;
         try {
             comentarios = comentarioDAO.getAll();
-        }catch (ExceptionDAO e){
-            throw new ExceptionService(e);
+        }catch (CustomException e){
+            throw e;
         }catch (Exception ex){
-            throw new ExceptionService(ex);
+            throw new CustomException(ex);
         }
         return comentarios;
     }
 
     @Override
-    public Comentario findOne(int idPublicacion) throws ExceptionService {
+    public Comentario findOne(int idPublicacion) throws CustomException {
         Comentario comentario;
         try {
             comentario = comentarioDAO.getByPublication(idPublicacion);
-        }catch (ExceptionDAO e){
-            throw new ExceptionService(e);
+            if(comentario == null) {
+                throw new CustomException(String.format(MessagesConstant.ENTITY_NOT_EXISTS,"Usuario"));
+            }
+        }catch (CustomException e){
+            throw e;
         }catch (Exception ex){
-            throw new ExceptionService(ex);
+            throw new CustomException(ex);
         }
         return comentario;
     }
 
     @Override
-    public void insertComentario(Comentario comentario) throws ExceptionService {
+    public void insertComentario(Comentario comentario) throws CustomException {
+        Comentario comentarioExistente = comentarioDAO.getByPublication(comentario.getIdPublicacion());
+        if (comentarioExistente != null){
+            throw new CustomException(String.format(MessagesConstant.ENTITY_ALREADY_EXISTS, "Comentario"));
+        }
         try {
             comentarioDAO.insert(comentario);
-        }catch (ExceptionDAO e){
-            throw new ExceptionService(e);
+        }catch (CustomException e){
+            throw e;
         }catch (Exception ex){
-            throw new ExceptionService(ex);
+            throw new CustomException(ex);
         }
     }
 
     @Override
-    public void updateComentario(Comentario comentario) throws ExceptionService {
+    public void updateComentario(Comentario comentario) throws CustomException {
+        Comentario comentarioExistente = comentarioDAO.getByPublication(comentario.getIdPublicacion());
+        if (comentarioExistente == null){
+            throw new CustomException(String.format(MessagesConstant.ENTITY_NOT_EXISTS, "Comentario"));
+        }
         try {
             comentarioDAO.update(comentario);
-        }catch (ExceptionDAO e){
-            throw new ExceptionService(e);
+        }catch (CustomException e){
+            throw e;
         }catch (Exception ex){
-            throw new ExceptionService(ex);
+            throw new CustomException(ex);
         }
     }
 
     @Override
-    public void deleteComentario(int idCometario) throws ExceptionService {
+    public void deleteComentario(int idCometario) throws CustomException {
+        Comentario comentarioExistente = comentarioDAO.getByPublication(idCometario);
+        if (comentarioExistente == null){
+            throw new CustomException(String.format(MessagesConstant.ENTITY_NOT_EXISTS, "Comentario"));
+        }
         try {
             comentarioDAO.delete(idCometario);
-        }catch (ExceptionDAO e){
-            throw new ExceptionService(e);
+        }catch (CustomException e){
+            throw e;
         }catch (Exception ex){
-            throw new ExceptionService(ex);
+            throw new CustomException(ex);
         }
     }
 }
